Split up tests in separate files
