+++ test/utils/web3.js
@@ -1,0 +1,5 @@
+import Web3 from 'web3';
+
+const web3 = new Web3();
+
+export default web3;
